refactor: PopulationChartContainerをcomponents/organismsに移動
